Updated runes stuff to work with bootbox.js (remove predefined modals, updated JS/views).

+++ herders/views.py
@@ -1366,23 +1366,13 @@ def rune_delete(request, profile_name, rune_id):
     summoner = get_object_or_404(Summoner, user__username=profile_name)
     is_owner = (request.user.is_authenticated() and summoner.user == request.user)
 
-    if is_owner:
-        if request.POST:
-            messages.success(request, 'Deleted ' + str(rune))
-            rune.delete()
+    if is_owner and request.method == 'POST':
+        messages.success(request, 'Deleted ' + str(rune))
+        rune.delete()
 
-            response_data = {
-                'code': 'success',
-            }
-        else:
-            form = DeleteRuneInstanceForm()
-            form.helper.form_action = reverse('herders:rune_delete', kwargs={'profile_name': profile_name, 'rune_id': rune_id})
-            template = loader.get_template('herders/profile/runes/delete_form.html')
-
-            response_data = {
-                'code': 'error',
-                'html': template.render(RequestContext(request, {'delete_rune_form': form}))
-            }
+        response_data = {
+            'code': 'success',
+        }
 
         return JsonResponse(response_data)
     else:
